requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.1.1
contextlib2==0.6.0.post1
cycler==0.10.0
cytoolz==0.10.1
dask==2.10.1
decorator==4.4.1
dm-sonnet==1.35
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.8.0
idna==2.8
imageio==2.6.1
Keras==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
networkx==2.4
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygpu==0.7.6
pyparsing==2.2.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scipy==1.3.2
semantic-version==2.8.4
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-probability==0.8.0
termcolor==1.1.0
Theano==1.0.4
toolz==0.10.0
tornado==6.0.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.11.2
